Cleanups.
(cherry picked from commit 333471df0130aa5d80259539ba8896e353f1db8f)
