+++ substratevm/src/com.oracle.svm.core.genscavenge/src/com/oracle/svm/core/genscavenge/AbstractCollectionPolicy.java
@@ -29,28 +29,18 @@ import static com.oracle.svm.core.Uninterruptible.CALLED_FROM_UNINTERRUPTIBLE_CO
 import org.graalvm.nativeimage.Platform;
 import org.graalvm.nativeimage.Platforms;
 import org.graalvm.nativeimage.StackValue;
-import org.graalvm.nativeimage.c.struct.RawField;
-import org.graalvm.nativeimage.c.struct.RawStructure;
-import org.graalvm.nativeimage.c.struct.SizeOf;
-import org.graalvm.word.Pointer;
-import org.graalvm.word.PointerBase;
 import org.graalvm.word.UnsignedWord;
 
 import com.oracle.svm.core.IsolateArgumentParser;
 import com.oracle.svm.core.SubstrateGCOptions;
 import com.oracle.svm.core.Uninterruptible;
-import com.oracle.svm.core.UnmanagedMemoryUtil;
 import com.oracle.svm.core.heap.PhysicalMemory;
-import com.oracle.svm.core.heap.ReferenceAccess;
 import com.oracle.svm.core.jdk.UninterruptibleUtils;
 import com.oracle.svm.core.locks.VMMutex;
-import com.oracle.svm.core.memory.NullableNativeMemory;
-import com.oracle.svm.core.nmt.NmtCategory;
 import com.oracle.svm.core.os.CommittedMemoryProvider;
 import com.oracle.svm.core.stack.StackOverflowCheck;
 import com.oracle.svm.core.thread.RecurringCallbackSupport;
 import com.oracle.svm.core.thread.VMOperation;
-import com.oracle.svm.core.thread.VMThreads;
 import com.oracle.svm.core.util.BasedOnJDKFile;
 import com.oracle.svm.core.util.UnsignedUtils;
 import com.oracle.svm.core.util.VMError;
@@ -63,7 +53,7 @@ import jdk.graal.compiler.word.Word;
  * to avoid inconsistencies between the different heap size values.
  */
 abstract class AbstractCollectionPolicy implements CollectionPolicy {
-    private static final VMMutex MUTEX = new VMMutex("AbstractCollectionPolicy.sizeParams");
+    private static final VMMutex SIZES_MUTEX = new VMMutex("AbstractCollectionPolicy.sizes");
 
     protected static final int MIN_SPACE_SIZE_AS_NUMBER_OF_ALIGNED_CHUNKS = 8;
 
@@ -167,11 +157,22 @@ abstract class AbstractCollectionPolicy implements CollectionPolicy {
         VMError.guarantee(sizes.isInitialized());
     }
 
+    /**
+     * (Re)computes and updates the {@link SizeParameters} used by this garbage collection policy to
+     * reflect the latest values derived from relevant heap size options and physical memory
+     * constraints.
+     * <p>
+     * This method is thread-safe and may be invoked concurrently by multiple threads. It is called
+     * at least once during VM startup, and is called again if the value of a relevant runtime
+     * option changes. A mutex is used to guarantee consistency of the computed parameters and
+     * prevent lost updates (note that a CAS would not be sufficient to prevent lost updates because
+     * threads could overwrite more recent values with outdated values).
+     */
     @Override
     public void updateSizeParameters() {
         StackOverflowCheck.singleton().makeYellowZoneAvailable();
         try {
-            MUTEX.lock();
+            SIZES_MUTEX.lock();
             try {
                 assert RecurringCallbackSupport.isCallbackUnsupportedOrTimerSuspended() : "recurring callbacks could trigger recursive locking, which isn't supported";
 
@@ -180,7 +181,7 @@ abstract class AbstractCollectionPolicy implements CollectionPolicy {
 
                 sizes.update(newValuesOnStack);
             } finally {
-                MUTEX.unlock();
+                SIZES_MUTEX.unlock();
             }
         } finally {
             StackOverflowCheck.singleton().protectYellowZone();
@@ -220,7 +221,7 @@ abstract class AbstractCollectionPolicy implements CollectionPolicy {
 
     @Override
     public final UnsignedWord getCurrentHeapCapacity() {
-        return sizes.getCurrentHeapCapacity();
+        return sizes.getHeapSize();
     }
 
     @Override
@@ -291,7 +292,7 @@ abstract class AbstractCollectionPolicy implements CollectionPolicy {
     @BasedOnJDKFile("https://github.com/openjdk/jdk/blob/jdk-25+21/src/hotspot/share/gc/shared/genArguments.cpp#L195-L310")
     @BasedOnJDKFile("https://github.com/openjdk/jdk/blob/jdk-25+21/src/hotspot/share/gc/parallel/psYoungGen.cpp#L104-L116")
     @BasedOnJDKFile("https://github.com/openjdk/jdk/blob/jdk-25+21/src/hotspot/share/gc/parallel/psYoungGen.cpp#L146-L168")
-    private void computeSizeParameters(RawSizeParameters newParams) {
+    private void computeSizeParameters(RawSizeParameters newParamsOnStack) {
         UnsignedWord minYoungSpaces = minSpaceSize(); // eden
         if (HeapParameters.getMaxSurvivorSpaces() > 0) {
             minYoungSpaces = minYoungSpaces.add(minSpaceSize().multiply(2)); // survivor from and to
@@ -392,7 +393,7 @@ abstract class AbstractCollectionPolicy implements CollectionPolicy {
         UnsignedWord youngSize = edenSize.add(survivorSize);
         UnsignedWord heapSize = edenSize.add(survivorSize).add(oldSize);
 
-        RawSizeParametersAccess.initialize(newParams,
+        RawSizeParametersOnStackAccess.initialize(newParamsOnStack,
                         initialEden, edenSize, maxEdenSize,
                         initialSurvivor, survivorSize, maxSurvivorSize,
                         initialOldSize, oldSize, maxOldSize,
@@ -426,432 +427,5 @@ abstract class AbstractCollectionPolicy implements CollectionPolicy {
 
     protected UnsignedWord getInitialHeapSize() {
         return AbstractCollectionPolicy.INITIAL_HEAP_SIZE;
-    }
-
-    /**
-     * Once a {@link RawSizeParameters} struct is visible to other threads (see {@link #update}), it
-     * may only be accessed by {@link Uninterruptible} code or when the VM is at a safepoint. This
-     * is necessary to prevent use-after-free errors because no longer needed
-     * {@link RawSizeParameters} structs may be freed during a GC.
-     * <p>
-     * When the VM is at a safepoint, the GC may directly update some of the values in the latest
-     * {@link RawSizeParameters} (see setters below).
-     */
-    protected static final class SizeParameters {
-        private static final String ACCESS_RAW_SIZE_PARAMETERS = "Prevent that RawSizeParameters are freed.";
-
-        private volatile RawSizeParameters sizes;
-
-        @Platforms(Platform.HOSTED_ONLY.class)
-        private SizeParameters() {
-        }
-
-        @Uninterruptible(reason = CALLED_FROM_UNINTERRUPTIBLE_CODE, mayBeInlined = true)
-        public boolean isInitialized() {
-            /*
-             * This only checks for non-null and doesn't access any values in the struct, so
-             * inlining into interruptible code is allowed.
-             */
-            return sizes.isNonNull();
-        }
-
-        @Uninterruptible(reason = ACCESS_RAW_SIZE_PARAMETERS)
-        public UnsignedWord getInitialEdenSize() {
-            assert isInitialized();
-            return sizes.getInitialEdenSize();
-        }
-
-        @Uninterruptible(reason = ACCESS_RAW_SIZE_PARAMETERS)
-        public UnsignedWord getEdenSize() {
-            assert isInitialized();
-            return sizes.getEdenSize();
-        }
-
-        @Uninterruptible(reason = ACCESS_RAW_SIZE_PARAMETERS)
-        public void setEdenSize(UnsignedWord value) {
-            assert isInitialized();
-            assert VMOperation.isGCInProgress();
-
-            sizes.setEdenSize(value);
-        }
-
-        @Uninterruptible(reason = ACCESS_RAW_SIZE_PARAMETERS)
-        public UnsignedWord getMaxEdenSize() {
-            assert isInitialized();
-            return sizes.getMaxEdenSize();
-        }
-
-        @Uninterruptible(reason = ACCESS_RAW_SIZE_PARAMETERS)
-        public UnsignedWord getInitialSurvivorSize() {
-            assert isInitialized();
-            return sizes.getInitialSurvivorSize();
-        }
-
-        @Uninterruptible(reason = ACCESS_RAW_SIZE_PARAMETERS)
-        public UnsignedWord getSurvivorSize() {
-            assert isInitialized();
-            return sizes.getSurvivorSize();
-        }
-
-        @Uninterruptible(reason = ACCESS_RAW_SIZE_PARAMETERS)
-        public void setSurvivorSize(UnsignedWord value) {
-            assert isInitialized();
-            assert VMOperation.isGCInProgress();
-
-            sizes.setSurvivorSize(value);
-        }
-
-        @Uninterruptible(reason = ACCESS_RAW_SIZE_PARAMETERS)
-        @BasedOnJDKFile("https://github.com/openjdk/jdk/blob/jdk-25+21/src/hotspot/share/gc/parallel/psYoungGen.cpp#L104-L116")
-        UnsignedWord getMaxSurvivorSize() {
-            assert isInitialized();
-            return sizes.getMaxSurvivorSize();
-        }
-
-        @Uninterruptible(reason = ACCESS_RAW_SIZE_PARAMETERS)
-        UnsignedWord getInitialYoungSize() {
-            assert isInitialized();
-            return sizes.getInitialYoungSize();
-        }
-
-        @Uninterruptible(reason = ACCESS_RAW_SIZE_PARAMETERS)
-        public UnsignedWord getYoungSize() {
-            assert isInitialized();
-            return sizes.getYoungSize();
-        }
-
-        @Uninterruptible(reason = ACCESS_RAW_SIZE_PARAMETERS)
-        public UnsignedWord getMaxYoungSize() {
-            assert isInitialized();
-            return sizes.getMaxYoungSize();
-        }
-
-        @Uninterruptible(reason = ACCESS_RAW_SIZE_PARAMETERS)
-        UnsignedWord getInitialOldSize() {
-            assert isInitialized();
-            return sizes.getInitialOldSize();
-        }
-
-        @Uninterruptible(reason = ACCESS_RAW_SIZE_PARAMETERS)
-        public UnsignedWord getOldSize() {
-            assert isInitialized();
-            return sizes.getOldSize();
-        }
-
-        @Uninterruptible(reason = ACCESS_RAW_SIZE_PARAMETERS)
-        public void setOldSize(UnsignedWord value) {
-            assert isInitialized();
-            assert VMOperation.isGCInProgress();
-
-            sizes.setOldSize(value);
-        }
-
-        @Uninterruptible(reason = ACCESS_RAW_SIZE_PARAMETERS)
-        UnsignedWord getMaxOldSize() {
-            assert isInitialized();
-            return sizes.getMaxOldSize();
-        }
-
-        @Uninterruptible(reason = ACCESS_RAW_SIZE_PARAMETERS)
-        public UnsignedWord getPromoSize() {
-            assert isInitialized();
-            return sizes.getPromoSize();
-        }
-
-        @Uninterruptible(reason = ACCESS_RAW_SIZE_PARAMETERS)
-        public void setPromoSize(UnsignedWord value) {
-            assert isInitialized();
-            assert VMOperation.isGCInProgress();
-
-            sizes.setPromoSize(value);
-        }
-
-        @Uninterruptible(reason = ACCESS_RAW_SIZE_PARAMETERS)
-        public UnsignedWord getMinHeapSize() {
-            assert isInitialized();
-            return sizes.getMinHeapSize();
-        }
-
-        @Uninterruptible(reason = ACCESS_RAW_SIZE_PARAMETERS)
-        public UnsignedWord getMaxHeapSize() {
-            assert isInitialized();
-            return sizes.getMaxHeapSize();
-        }
-
-        @Uninterruptible(reason = ACCESS_RAW_SIZE_PARAMETERS)
-        public UnsignedWord getCurrentHeapCapacity() {
-            assert isInitialized();
-            assert VMOperation.isGCInProgress() : "use only during GC";
-
-            return sizes.getHeapSize();
-        }
-
-        @Uninterruptible(reason = ACCESS_RAW_SIZE_PARAMETERS)
-        private void update(RawSizeParameters newValuesOnStack) {
-            RawSizeParameters prevValues = sizes;
-            if (prevValues.isNonNull() && matches(prevValues, newValuesOnStack)) {
-                /* Nothing to do - cached params are still accurate. */
-                return;
-            }
-
-            /* Try allocating a struct on the C heap. */
-            UnsignedWord structSize = SizeOf.unsigned(RawSizeParameters.class);
-            RawSizeParameters newValuesOnHeap = NullableNativeMemory.malloc(structSize, NmtCategory.GC);
-            VMError.guarantee(newValuesOnHeap.isNonNull(), "Out-of-memory while updating GC policy sizes.");
-
-            /* Copy the values from the stack to the C heap. */
-            UnmanagedMemoryUtil.copyForward((Pointer) newValuesOnStack, (Pointer) newValuesOnHeap, structSize);
-            newValuesOnHeap.setNext(prevValues);
-
-            /*
-             * Publish the new struct via a volatile store. Once the data is published, other
-             * threads need to see a fully initialized struct right away. This is guaranteed by the
-             * implicit STORE_STORE barrier before the volatile write.
-             */
-            sizes = newValuesOnHeap;
-
-            assert isAligned(getMaxSurvivorSize()) && isAligned(getInitialYoungSize()) && isAligned(getInitialOldSize()) && isAligned(getMaxOldSize());
-            assert getMaxSurvivorSize().belowThan(getMaxYoungSize());
-            assert getMaxYoungSize().add(getMaxOldSize()).equal(getMaxHeapSize());
-            assert getInitialEdenSize().add(getInitialSurvivorSize().multiply(2)).equal(getInitialYoungSize());
-            assert getInitialYoungSize().add(getInitialOldSize()).equal(sizes.getInitialHeapSize());
-        }
-
-        /**
-         * Frees no longer needed {@link RawSizeParameters}, so that only the most recent one
-         * remains.
-         */
-        public void freeUnusedSizeParameters() {
-            assert isInitialized();
-            assert VMOperation.isGCInProgress() : "would need to be uninterruptible otherwise";
-
-            RawSizeParameters cur = sizes;
-            freeSizeParameters(cur.getNext());
-            cur.setNext(Word.nullPointer());
-        }
-
-        @Uninterruptible(reason = "Tear-down in progress.")
-        public void tearDown() {
-            assert VMThreads.isTearingDown();
-
-            freeSizeParameters(sizes);
-            sizes = Word.nullPointer();
-        }
-
-        @Uninterruptible(reason = CALLED_FROM_UNINTERRUPTIBLE_CODE, mayBeInlined = true)
-        private static void freeSizeParameters(RawSizeParameters first) {
-            assert VMOperation.isGCInProgress() || VMThreads.isTearingDown();
-
-            RawSizeParameters cur = first;
-            while (cur.isNonNull()) {
-                RawSizeParameters next = cur.getNext();
-                NullableNativeMemory.free(cur);
-                cur = next;
-            }
-        }
-
-        @Uninterruptible(reason = ACCESS_RAW_SIZE_PARAMETERS)
-        private static boolean matches(RawSizeParameters a, RawSizeParameters b) {
-            return a.getInitialEdenSize() == b.getInitialEdenSize() &&
-                            a.getEdenSize() == b.getEdenSize() &&
-                            a.getMaxEdenSize() == b.getMaxEdenSize() &&
-
-                            a.getInitialSurvivorSize() == b.getInitialSurvivorSize() &&
-                            a.getSurvivorSize() == b.getSurvivorSize() &&
-                            a.getMaxSurvivorSize() == b.getMaxSurvivorSize() &&
-
-                            a.getInitialYoungSize() == b.getInitialYoungSize() &&
-                            a.getYoungSize() == b.getYoungSize() &&
-                            a.getMaxYoungSize() == b.getMaxYoungSize() &&
-
-                            a.getInitialOldSize() == b.getInitialOldSize() &&
-                            a.getOldSize() == b.getOldSize() &&
-                            a.getMaxOldSize() == b.getMaxOldSize() &&
-
-                            a.getPromoSize() == b.getPromoSize() &&
-
-                            a.getMinHeapSize() == b.getMinHeapSize() &&
-                            a.getInitialHeapSize() == b.getInitialHeapSize() &&
-                            a.getHeapSize() == b.getHeapSize() &&
-                            a.getMaxHeapSize() == b.getMaxHeapSize();
-        }
-    }
-
-    /**
-     * Struct that stores all GC-related sizes. See {@link SizeParameters} for more details on its
-     * lifecycle.
-     */
-    @RawStructure
-    private interface RawSizeParameters extends PointerBase {
-        @RawField
-        UnsignedWord getInitialEdenSize();
-
-        @RawField
-        void setInitialEdenSize(UnsignedWord value);
-
-        @RawField
-        UnsignedWord getEdenSize();
-
-        @RawField
-        void setEdenSize(UnsignedWord value);
-
-        @RawField
-        UnsignedWord getMaxEdenSize();
-
-        @RawField
-        void setMaxEdenSize(UnsignedWord value);
-
-        @RawField
-        UnsignedWord getInitialSurvivorSize();
-
-        @RawField
-        void setInitialSurvivorSize(UnsignedWord value);
-
-        @RawField
-        UnsignedWord getSurvivorSize();
-
-        @RawField
-        void setSurvivorSize(UnsignedWord value);
-
-        @RawField
-        UnsignedWord getMaxSurvivorSize();
-
-        @RawField
-        void setMaxSurvivorSize(UnsignedWord value);
-
-        @RawField
-        UnsignedWord getInitialYoungSize();
-
-        @RawField
-        void setInitialYoungSize(UnsignedWord value);
-
-        @RawField
-        UnsignedWord getYoungSize();
-
-        @RawField
-        void setYoungSize(UnsignedWord value);
-
-        @RawField
-        UnsignedWord getMaxYoungSize();
-
-        @RawField
-        void setMaxYoungSize(UnsignedWord value);
-
-        @RawField
-        UnsignedWord getInitialOldSize();
-
-        @RawField
-        void setInitialOldSize(UnsignedWord value);
-
-        @RawField
-        UnsignedWord getOldSize();
-
-        @RawField
-        void setOldSize(UnsignedWord value);
-
-        @RawField
-        UnsignedWord getMaxOldSize();
-
-        @RawField
-        void setMaxOldSize(UnsignedWord value);
-
-        @RawField
-        UnsignedWord getPromoSize();
-
-        @RawField
-        void setPromoSize(UnsignedWord value);
-
-        @RawField
-        UnsignedWord getMinHeapSize();
-
-        @RawField
-        void setMinHeapSize(UnsignedWord value);
-
-        @RawField
-        UnsignedWord getInitialHeapSize();
-
-        @RawField
-        void setInitialHeapSize(UnsignedWord value);
-
-        @RawField
-        UnsignedWord getHeapSize();
-
-        @RawField
-        void setHeapSize(UnsignedWord value);
-
-        @RawField
-        UnsignedWord getMaxHeapSize();
-
-        @RawField
-        void setMaxHeapSize(UnsignedWord value);
-
-        @RawField
-        RawSizeParameters getNext();
-
-        @RawField
-        void setNext(RawSizeParameters value);
-    }
-
-    /**
-     * The methods in this class may only be used on {@link RawSizeParameters} that are allocated on
-     * the stack and therefore not yet visible to other threads. Please keep the methods in this
-     * class to a minimum.
-     */
-    private static final class RawSizeParametersAccess {
-        static void initialize(RawSizeParameters valuesOnStack,
-                        UnsignedWord initialEdenSize, UnsignedWord edenSize, UnsignedWord maxEdenSize,
-                        UnsignedWord initialSurvivorSize, UnsignedWord survivorSize, UnsignedWord maxSurvivorSize,
-                        UnsignedWord initialOldSize, UnsignedWord oldSize, UnsignedWord maxOldSize,
-                        UnsignedWord promoSize,
-                        UnsignedWord initialYoungSize, UnsignedWord youngSize, UnsignedWord maxYoungSize,
-                        UnsignedWord minHeapSize, UnsignedWord initialHeapSize, UnsignedWord heapSize, UnsignedWord maxHeapSize) {
-            assert isAligned(maxHeapSize) && isAligned(maxYoungSize) && isAligned(initialHeapSize) && isAligned(initialEdenSize) && isAligned(initialSurvivorSize);
-
-            assert initialEdenSize.belowOrEqual(initialYoungSize);
-            assert edenSize.belowOrEqual(youngSize);
-            assert maxEdenSize.belowOrEqual(maxYoungSize);
-
-            assert initialSurvivorSize.belowOrEqual(initialYoungSize);
-            assert survivorSize.belowOrEqual(youngSize);
-            assert maxSurvivorSize.belowOrEqual(maxYoungSize);
-
-            assert initialOldSize.belowOrEqual(initialHeapSize);
-            assert oldSize.belowOrEqual(heapSize);
-            assert maxOldSize.belowOrEqual(maxHeapSize);
-
-            assert initialYoungSize.belowOrEqual(initialHeapSize);
-            assert youngSize.belowOrEqual(heapSize);
-            assert maxYoungSize.belowOrEqual(maxHeapSize);
-
-            assert minHeapSize.belowOrEqual(initialHeapSize);
-            assert initialHeapSize.belowOrEqual(maxHeapSize);
-            assert heapSize.belowOrEqual(maxHeapSize);
-            assert maxHeapSize.belowOrEqual(ReferenceAccess.singleton().getMaxAddressSpaceSize());
-
-            valuesOnStack.setInitialEdenSize(initialEdenSize);
-            valuesOnStack.setEdenSize(edenSize);
-            valuesOnStack.setMaxEdenSize(maxEdenSize);
-
-            valuesOnStack.setInitialSurvivorSize(initialSurvivorSize);
-            valuesOnStack.setSurvivorSize(survivorSize);
-            valuesOnStack.setMaxSurvivorSize(maxSurvivorSize);
-
-            valuesOnStack.setInitialYoungSize(initialYoungSize);
-            valuesOnStack.setYoungSize(youngSize);
-            valuesOnStack.setMaxYoungSize(maxYoungSize);
-
-            valuesOnStack.setInitialOldSize(initialOldSize);
-            valuesOnStack.setOldSize(oldSize);
-            valuesOnStack.setMaxOldSize(maxOldSize);
-
-            valuesOnStack.setPromoSize(promoSize);
-
-            valuesOnStack.setMinHeapSize(minHeapSize);
-            valuesOnStack.setInitialHeapSize(initialHeapSize);
-            valuesOnStack.setHeapSize(heapSize);
-            valuesOnStack.setMaxHeapSize(maxHeapSize);
-
-            valuesOnStack.setNext(Word.nullPointer());
-        }
     }
 }
